+++ tests/conftest.py
@@ -197,6 +197,22 @@ class RunResult:
 
         return result
 
+    @property
+    def errors(self):
+        result = []
+        record = False
+        for line in self._result.stdout.lines:
+            line = line.strip()
+
+            if line.startswith("====") and "ERRORS" in line:
+                record = True
+            if record and line:
+                result.append(line)
+        result = self._join_lines(result)
+
+        result = re.sub(r"\d+\.\d+s", "<time>", result)
+        return result
+
     def errorLines(self):
         return self._join_lines(
             [line for line in self.stdout.lines if line and line[:2] in ("> ", "E ")]
@@ -206,6 +222,10 @@ class RunResult:
 @pytest.fixture
 def project(pytester):
     class Project:
+
+        def __init__(self):
+            self.term_columns = 80
+
         def setup(self, source: str):
             self.header = """\
 # äöß 🐍
@@ -283,12 +303,15 @@ def set_time(time_machine):
                 del os.environ["CI"]  # pragma: no cover
 
             try:
-                # result = pytester.runpytest_subprocess(*args)
                 with mock.patch.dict(
                     os.environ,
                     {
                         "TERM": "unknown",
-                        "COLUMNS": "81" if platform.system() == "Windows" else "80",
+                        "COLUMNS": str(
+                            self.term_columns + 1
+                            if platform.system() == "Windows"
+                            else self.term_columns
+                        ),
                     },
                 ):